% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{density.a.b.exact}
\alias{density.a.b.exact}
\title{Non-critical clone size distribution (exact).}
\usage{
\method{density}{a.b.exact}(lambda, delta, t, a, b)
}
\arguments{
\item{lambda}{proliferation rate}

\item{delta}{loss rate}

\item{t}{time}

\item{a}{clone size at t=0}

\item{b}{clone size at t=t}
}
\value{
The probability that a clone of size a grows to size b within t.
}
\description{
Exact probability to grow from a clone of size a to a clone of size b within time t according to a non-critical birth-death process. Exact solution according to Bailey, 1964.
}
\examples{
density.a.b.exact(1, 0, 10, 1, 2)
}
