% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_fasta.R
\name{df_to_fasta}
\alias{df_to_fasta}
\title{Fasta file creation from dataframe columns and/or vectors.}
\usage{
df_to_fasta(
  sequence_name,
  sequence_strings,
  file_name = "sequences.fasta",
  output_dir = NULL
)
}
\arguments{
\item{sequence_name}{Vector containing the names for each sequence, usually a column from a `data.frame`. eg. df$sequence_name}

\item{sequence_strings}{Vector containing the sequences (DNA, RNA, AA), usually a column from a `data.frame`. eg. df$sequences}

\item{file_name}{Output file name to be saved as a fasta file.}

\item{output_dir}{Output directory for the fasta file. Default is the working directory.}
}
\value{
Saves a fasta file in the desired location, and also returns the stringset as BStringSet if saved as an object.
}
\description{
Fasta file creation from dataframe columns and/or vectors.
}
\examples{
\dontrun{
df_to_fasta(sequence_name = df$sequence_name,
sequence_strings = df$sequence_name,
file_name = "my_fasta.fasta",
output_dir = "path/to/output")
}
}
