% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{.alpha}
\alias{.alpha}
\title{Exact probability to grow from a to be within time t according to a supercritical birth-death process. Exact solution according to Bailey, 1964}
\usage{
.alpha(lambda, delta, t)
}
\arguments{
\item{lambda}{proliferation rate}

\item{delta}{loss rate}

\item{t}{time}

\item{a}{clone size at t=0}

\item{b}{clone size at t=t}
}
\value{
The probability that a clone of size a grows to size b within t.
}
\description{
Exact probability to grow from a to be within time t according to a supercritical birth-death process. Exact solution according to Bailey, 1964
}
\examples{
density.a.b.exact(1, 0, 10, 1, 2)
}
