% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{.approximate.delta}
\alias{.approximate.delta}
\title{Approximate deltas to approximate non-linear decline with a linear b-d-process parametrized by the number of death events}
\usage{
.approximate.delta(lambda, N, t, D)
}
\arguments{
\item{lambda}{the division rate}

\item{N}{the clone size at the start of contraction}

\item{t}{the time span}

\item{D}{the number of death events}
}
\value{
the death rate yielding the same number of death events if modeling exponential decay
}
\description{
Approximate deltas to approximate non-linear decline with a linear b-d-process parametrized by the number of death events
}
