% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{.clonal.combinations}
\alias{.clonal.combinations}
\title{Clonal combinations}
\usage{
.clonal.combinations(clone.ids)
}
\arguments{
\item{clone.ids}{IDs of daughter clones from the same mother}
}
\value{
All clonal combinations in which a mutation acquired in the mother can end up.
}
\description{
Clonal combinations
}
