% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution (conflicted copy
%   2023-06-16 181130).R
\name{.clonal_dynamics}
\alias{.clonal_dynamics}
\title{Dynamics of normal cells and i selected clones in a hierarchical tissue with j compartments}
\usage{
.clonal_dynamics(N, init, lambda, delta, s, t)
}
\arguments{
\item{N}{vector of length j with carrying capacities in each compartment}

\item{init}{ixj with the initial condition of the system (i: compartment, j: clone)}

\item{lambda}{vector of length j with cell division rates in each compartment}

\item{delta}{vector of differentiation rates in each compartment}

\item{s}{ixj matrix with selective advantage associated with the i-th driver in the j-th compartment. Selection is modeled as a reduction of the differentiation rate, so 0 <= s <= 1}

\item{t}{the time point of evaluation}
}
\value{
The system state at time t
}
\description{
Dynamics of normal cells and i selected clones in a hierarchical tissue with j compartments
}
