% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{.compute_actual_size}
\alias{.compute_actual_size}
\title{Compute the final size of the clones from the input parameters}
\usage{
.compute_actual_size(
  t.s,
  mother.daughter,
  N,
  lambda.ss,
  delta.ss,
  lambda.exp,
  delta.exp,
  size,
  t.end
)
}
\arguments{
\item{t.s}{vector of length j with the time points at which the selected advantages were introduced}

\item{mother.daughter}{mother-daughter relationships. Matrix with 2 columns encoding mother and daughter for each pair}

\item{N}{the compartment size}

\item{lambda.ss}{cell division rate during homeostasis}

\item{delta.ss}{differentiation rate during homeostasis}

\item{lambda.exp}{division rate during initial expansion}

\item{delta.exp}{loss rate during initial expansion}

\item{size}{vector of length j with the input size of the selcted clones. Using exponential growth approximation, the selective advantage, \code{s} will be computed from \code{size}. The actual clone sizes will then be computed using the values of \code{s} and a clonal competition model.}

\item{t.end}{the time point of evaluation}
}
\value{
A vector reporting the system state at the final time point for each clone
}
\description{
Compute the final size of the clones from the input parameters
}
