% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution (conflicted copy
%   2023-06-16 181130).R
\name{.deltas}
\alias{.deltas}
\title{Determine the loss rates from logistic growth}
\usage{
.deltas(delta, s, N, state)
}
\arguments{
\item{delta}{a vector of length i containing the loss rate per compartment}

\item{s}{ixj matrix with selective advantage associated with clone j in compartment i}

\item{N}{a vector of length i containing the compartment sizes}

\item{state}{ixj matrix with the number of cells of the j-th clone in compartment i}
}
\value{
a vector with the current loss rates
}
\description{
If there are driver mutations in the simulation, this function uses a competition model between drivers and normal cells. The corresponding ODE model reads (N: normal cells, D1: driver 1, D2: driver2; N + D1 + D2=K = const.):
\enumerate{
   \item{dN/dt  = lambda*N - rho_NN N^2/K - rho_ND1 N*D1/K - rho_ND2 N*D2/K and so on}
   \item{dD1/dt = lambda*D1 - rho_D1N D1*N/K - rho_D1D1 * D1^2/K - rho_D1D2 D1*D2/K and so on}
   \item{dD2/dt = lambda*D2 - rho*D2N D2*N/K - rho_D2D1 * D1*D2/K - rho_D2D2 * D2^2/K and so on}
}
We require
\enumerate{
\item rho_NN=rho_D1D1=rho_D2D2=delta/K
\item rho_D1N=r1/K, rhoD2N=r2/K, rhoD2D1=r2/r1/K
\item rho_ND1=2-r1, rho_ND2=2-r2, rhoD1D2=2-r2/r1 and so on
}
}
\keyword{internal}
