% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_trees.R
\name{.division.s.p}
\alias{.division.s.p}
\title{Cell division}
\usage{
.division.s.p(
  cell.type,
  tree,
  mut.rate.D = 0,
  s.shape = 1.5,
  s.rate = 35,
  mutation.mode,
  mut.rate,
  symmetric = T,
  nr = 1
)
}
\arguments{
\item{cell.type}{the cell type that is to divide (1, stem cell; 2, progenitor cell)}

\item{tree}{object of class phylo; the current tree. Needs the following additional list elements: tip.class, specifying the cell type and sel.adv., specifying the selective advantage of each tip}

\item{mut.rate.D}{integer; driver mutation rate per cell division}

\item{s.shape}{shape parameter of the gamma distribution from which the selective advantage of a new driver is drawn}

\item{s.rate}{rate parameter of the gamma distribution from which the selective advantage of a new driver is drawn}

\item{mutation.mode}{should the number of new mutations be "constant" or "Binomial"ly distributed?}

\item{mut.rate}{average number of neutral mutations per daughter cell and division}

\item{symmetric}{is the division a symmetric division (2 new daughter cells) or an asymmetric division (1 daughter cell differentiates); defaults to T}

\item{nr}{number of reactions to simulate; defaults to 1}
}
\value{
the updated tree
}
\description{
Simulate division of a cell together with acquisition of neutral and driver mutations
}
