% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{.forward_dynamics}
\alias{.forward_dynamics}
\title{Forward dynamics of j clones}
\usage{
.forward_dynamics(
  N,
  init.cells,
  lambda.ss,
  delta.ss,
  lambda.exp,
  delta.exp,
  s,
  t.s,
  mother.daughter,
  t,
  resolution = 0.01
)
}
\arguments{
\item{N}{the compartment size}

\item{init.cells}{vector with the initial condition of the system (number of cells per clone)}

\item{lambda.ss}{cell division rate during homeostasis}

\item{delta.ss}{differentiation rate during homeostasis}

\item{lambda.exp}{division rate during initial expansion}

\item{delta.exp}{loss rate during initial expansion}

\item{s}{vector with selective advantages associated with the j-th driver. Selection is modeled as a reduction of the differentiation rate, so 0 <= s <= 1}

\item{t.s}{vector of length j with the time points at which the selected advantages were introduced}

\item{mother.daughter}{mother-daughter relationships. Matrix with 2 columns encoding mother and daughter for each pair}

\item{t}{the time point of evaluation}

\item{resolution}{the time resolution of the simulation}
}
\value{
A data.frame reporting the system state in the following order: time, cell count for each clone
}
\description{
Forward dynamics of j clones
}
