% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{.get.progeny}
\alias{.get.progeny}
\title{Get all subclonal descendants from a given clone}
\usage{
.get.progeny(mother.daughter, id)
}
\arguments{
\item{mother.daughter}{mother-daughter relationships. Matrix with 3 columns encoding mother, daughter and birth-compartment for each pair}

\item{id}{IDs of the clone of interest}
}
\value{
A vector with the progeny IDs.
}
\description{
Get all subclonal descendants from a given clone
}
