% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_trees.R
\name{.initialize.sim.s.p}
\alias{.initialize.sim.s.p}
\title{System initialization}
\usage{
.initialize.sim.s.p()
}
\arguments{
\item{N}{integer; number of stem cells at homeostasis}

\item{NP}{integer; number of progenitor cells at homeostasis}

\item{mut.rate}{integer; number of mutations per cell division and daugther cell}

\item{time.max}{time of simulation}

\item{parms.steady}{named vector of steady state parameters; must contain lambda.s/lambda.p the stem and progenitor division rate; delta.s/delta.p the stem and progenitor loss rate and alpha.s the stem cell differentiation rate}
}
\value{
returns the state list; initialized with a single stem cell and a single mutation
}
\description{
This function initializes the simulation
}
