% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_trees.R
\name{.loss}
\alias{.loss}
\title{Cell loss}
\usage{
.loss(cell.type, tree, nr = 1)
}
\arguments{
\item{cell.type}{integer; which cell type is dividing? 1, stem cell, 2, progenitor cell}

\item{tree}{object of class phylo; the current tree. Needs the following additional list elements: \code{tip.class}, specifying the cell type and \code{sel.adv.}, specifying the selective advantage of each tip}

\item{nr}{number of reactions to simulate; defaults to 1}
}
\value{
the updated tree
}
\description{
Simulate loss of a stem cell or a progenitor cell
}
