% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{.loss_rates}
\alias{.loss_rates}
\title{Determine loss rates from logistic growth}
\usage{
.loss_rates(delta, s, N, cell.state)
}
\arguments{
\item{delta}{a vector of length i containing the loss rate per compartment}

\item{s}{a vector of selective advantages associated with clone j}

\item{N}{the compartment size}

\item{cell.state}{vector of length j with the number of cells of the j-th clone}
}
\value{
an vector of length j with the current loss rates
}
\description{
We assume that different clones compete with each other and that competition acts on the loss rate. This function computes the net loss rates for a given state in dependence of the selective advantages, according to
\enumerate{
\item rho_D0D0=rho_D1D1=rho_D2D2=...=rho_DiDi=delta/K
\item rho_D1D0=r1/K, rhoD2D0=r2/K, rhoD2D1=r2/r1/K, ...
\item rho_D0D1=2-r1, rho_D0D2=2-r2, rhoD1D2=2-r2/r1, ...
},
where \code{D}stands for \code{D}river
}
\keyword{internal}
