% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_trees.R
\name{.mutation.acquisition}
\alias{.mutation.acquisition}
\title{Acquisition of neutral mutations}
\usage{
.mutation.acquisition(mut.rate, mutation.mode = "Binomial")
}
\arguments{
\item{mut.rate}{integer; average number of mutations per division and daughter cell}

\item{mutation.mode}{character vector; should the number be based on a binomial distribution ("Binomial") or should a constant number be introduced ("constant"); defaults to "Binomial".}
}
\value{
the number of new mutations per daughter cell
}
\description{
This function simulates the acquisition of new mutations
}
