% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_trees.R
\name{.props.2c}
\alias{.props.2c}
\title{Propensity function for 2 compartments ("S"tem cells and "P"rogenitors)}
\usage{
.props.2c(cell.count, parms)
}
\arguments{
\item{cell.count}{a vector with cell numbers (S, P)}

\item{parms}{a named parameter vector containing the rates of cell division (lambda.s/lambda.p) and differentiation (alpha.s) or loss (delta.s/delta.p), where "s" and "p" indicate stem and progenitor cells, respectively}
}
\value{
the propensities of the possible reactions
}
\description{
This function computes the propensities for a linear, hierarchical system of cell division, death and differentiation
}
