% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_heterogenous_tissues.R
\name{.props.het}
\alias{.props.het}
\title{Propensity function for "S"tem cells and 2 different mature cells (Type 1 and Type 2)}
\usage{
.props.het(cell.count, parms)
}
\arguments{
\item{cell.count}{a vector with cell numbers (S, T1, T2)}

\item{parms}{a named parameter vector containing the rates of cell division (lambda.s/lambda.t1/lambda.t2) and differentiation (alpha.s1/alpha.s2) or loss (delta.s/delta.t1/delta.t2), where "s" and "t" indicate stem and type (1 or 2) cells, respectively}
}
\value{
the propensities of the possible reactions
}
\description{
This function computes the propensities for a linear, hierarchical system of cell division, death and differentiation
}
