% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_post_processing.R
\name{.simulate.sampling}
\alias{.simulate.sampling}
\title{Sample a random subset of a phylogenetic tree}
\usage{
.simulate.sampling(tree, sample.size)
}
\arguments{
\item{tree}{object of class \code{phylo}; the current tree. Needs the following additional list elements: \code{tip.class}, specifying the cell type and \code{sel.adv.}, specifying the selective advantage of each tip.}

\item{sample.size}{the number of cells to be sampled}
}
\value{
the updated state list
}
\description{
Sample a random subset of a phylogenetic tree
}
