% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution (conflicted copy
%   2023-06-16 181130).R, R/Clone_size_distribution.R
\name{.simulated.scWGS.data}
\alias{.simulated.scWGS.data}
\title{Simulate sampling as in single-cell sequencing by simulating the actual sampling of cells (i.e. it's not the expected value but a sampling instance)}
\usage{
.simulated.scWGS.data(
  clone.sizes,
  expected.mutations,
  ncells = 100,
  min.vaf = 0.05
)

.simulated.scWGS.data(
  clone.sizes,
  expected.mutations,
  ncells = 100,
  min.vaf = 0.05
)
}
\arguments{
\item{clone.sizes}{vector of clone sizes at which cumulative mutation counts were measured}

\item{expected.mutations}{expected number of mutations at each clone size}

\item{ncells}{sequenced cells}
}
\value{
A vector of simulated VAFs

A vector of simulated VAFs
}
\description{
Simulate sampling as in single-cell sequencing by simulating the actual sampling of cells (i.e. it's not the expected value but a sampling instance)

Simulate sampling as in single-cell sequencing by simulating the actual sampling of cells (i.e. it's not the expected value but a sampling instance)
}
