% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_post_processing.R
\name{.subset.cell.type}
\alias{.subset.cell.type}
\title{Subset a phylogenetic tree on a particular cell type}
\usage{
.subset.cell.type(tree, cell.type)
}
\arguments{
\item{tree}{object of class \code{phylo}; the current tree. Needs the following additional list elements: \code{tip.class}, specifying the cell type and \code{sel.adv.}, specifying the selective advantage of each tip.}

\item{cell.type}{cell type of interest}
}
\value{
the updated stat list
}
\description{
Subset a phylogenetic tree on a particular cell type
}
