% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution (conflicted copy
%   2023-06-16 181130).R
\name{ExpectedNumberAfterWGS}
\alias{ExpectedNumberAfterWGS}
\title{Compute the expected number of mutations after WGS (note: runs slowly!)}
\usage{
ExpectedNumberAfterWGS(M, vafs, depth, false.negative.per.vaf, min.vaf = 0.05)
}
\arguments{
\item{M}{vector of cumulative mutation counts with VAF as given in 'vaf'}

\item{depth}{sequencing depth}

\item{false.negative.per.vaf}{estimated fraction of false negatives per vaf; columns correspond to VAFs as in vaf, rows to individual measurements}

\item{vaf}{vector of variant allele frequencies for which M is reported}
}
\value{
A vector of simulated VAFs
}
\description{
Compute the expected number of mutations after WGS (note: runs slowly!)
}
