% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.quality.R
\name{get.processors}
\alias{get.processors}
\title{B cell receptor immune repertoire quality control and analysis}
\usage{
get.processors(processors)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Factor variable.
}
\description{
RepertoiR was created to do quality control of VDJ sequencing, combining single-cell sorted data and specificity with sanguer sequences.
}
\examples{
sf <- summarise.abi.folder("abi/folder/path", secondary.peak.ratio = 0.33, trim.cutoff = 0.01, processors = 1)

}
