% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_post_processing.R
\name{get_matrix_from_tree}
\alias{get_matrix_from_tree}
\title{Get a binary mutation matrix from a phylogenetic tree}
\usage{
get_matrix_from_tree(tree)
}
\arguments{
\item{tree}{object of class \code{phylo}; the current tree. Needs the following additional list elements: \code{tip.class}, specifying the cell type and \code{sel.adv.}, specifying the selective advantage of each tip.}
}
\value{
a matrix with rows corresponding to mutations and columns to cells; entries are binaries of 0 or 1, indicating, respectively, absence or presence of the mutation.
}
\description{
Get a binary mutation matrix from a phylogenetic tree
}
