% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_post_processing.R
\name{get_mutations_per_tip}
\alias{get_mutations_per_tip}
\title{Compute the number of mutations per tip cell in a phylogenetic tree}
\usage{
get_mutations_per_tip(tree)
}
\arguments{
\item{tree}{object of class \code{phylo}; the current tree. Needs the following additional list elements: \code{tip.class}, specifying the cell type and \code{sel.adv.}, specifying the selective advantage of each tip.}
}
\value{
a vector of mutation counts
}
\description{
Compute the number of mutations per tip cell in a phylogenetic tree
}
