% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_post_processing.R
\name{get_vaf_from_tree}
\alias{get_vaf_from_tree}
\title{Compute the VAF of a mutation in the population from a phylogenetic tree}
\usage{
get_vaf_from_tree(tree)
}
\arguments{
\item{tree}{object of class phylo; the current tree. Needs the following additional list elements: \code{tip.class}, specifying the cell type and \code{sel.adv.}, specifying the selective advantage of each tip.}
}
\value{
a vector of VAFs
}
\description{
Compute the VAF of a mutation in the population from a phylogenetic tree
}
