% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_trees.R
\name{gillespie.sim.s.p}
\alias{gillespie.sim.s.p}
\title{Tree simulation}
\usage{
gillespie.sim.s.p(
  parms.exp,
  parms.steady,
  time.max = 50,
  time.samples = c(0, 5, 25, 50),
  N = 1000,
  NP = 10000,
  report.at.f = NA,
  mut.rate = 3,
  mutation.mode = "Binomial",
  driver.mode = "random",
  t.driver = NA,
  mut.rate.D = 0,
  s.shape = 1.5,
  s.rate = 35,
  tau = 1
)
}
\arguments{
\item{parms.exp}{expansion parameters, vector that must contain \code{lambda.s} (stem cell division rate), \code{lambda.p} (progenitor cell division rate), \code{alpha.s} (stem cell differentiation rate), \code{delta.s} (stem cell loss rate), \code{delta.p} (progenitor cell differentiation rate)}

\item{parms.steady}{same as \code{parms.exp}, but for the homeostatic phase}

\item{time.max}{maximal simulation time}

\item{time.samples}{time points at which simulation results are stored}

\item{N}{number of stem cells during homeostasis}

\item{NP}{number of progenitor cells during homeostasis}

\item{report.at.f}{frequency of a selected clone at which the simulation should be stopped. Defaults to NA - don't stop at a certain frequency.}

\item{mut.rate}{average number of mutation per division and daughter cell}

\item{mutation.mode}{should mutations be "constant" or "Binomial"ly distributed?}

\item{driver.mode}{string; "fixed_time" if driver is acquired at a fixed time point, "random", if driver is randomly acquired governed by the parameters \code{mu_D} and \code{s}}

\item{t.driver}{integer, the time point at which the driver is acquired if mode is "fixed_time"}

\item{mut.rate.D}{integer; driver mutation rate per cell division}

\item{s.shape}{shape parameter of the gamma distribution from which the selective advantage of a new driver is drawn. If the function is run with \code{driver.mode="fixed"}, the selective advantage is not randomyl drawn but computed as s.shape/d.rate.}

\item{s.rate}{rate parameter of the gamma distribution from which the selective advantage of a new driver is drawn}

\item{tau}{tau leaping parameter: how many steps should be merged during homeostasis? Defaults to 1, no leaping.}
}
\value{
a list of state.lists at the desired time samples
}
\description{
Simulate the phylogenetic tree of a physiological population that grows in 2 regimes: initial exponential expansion and subsequent homeostasis. The function can simuate stem cells only or stem and progenitor cells
}
