% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_heterogenous_tissues.R
\name{gillespie.sim.s.t1.t2}
\alias{gillespie.sim.s.t1.t2}
\title{Tree simulation}
\usage{
gillespie.sim.s.t1.t2(
  parms.exp,
  parms.steady,
  time.max = 50,
  time.samples = c(0, 5, 25, 50),
  N = 1000,
  NT1 = 8000,
  NT2 = 2000,
  report.at.f = NA,
  mut.rate = 3,
  mutation.mode = "Binomial",
  driver.mode = "random",
  t.driver = NA,
  mother = NULL,
  mut.rate.D = 0,
  s.shape = 1.5,
  s.rate = 35,
  tau = 1
)
}
\arguments{
\item{parms.exp}{expansion parameters, vector that must contain lambda.s (stem cell division rate), lambda.t1 (type 1 cell division rate), lambda.t2 (type 2 cell division rate), alpha.s1 (stem cell differentiation rate into type 1 cells), alpha.s2 (stem cell division rate into type 2 cells), delta.s (stem cell loss rate), delta.t1 (type 1 cell differentiation rate), delta.t2 (type 2 cell differentiation rate)}

\item{parms.steady}{as parms.exp, but for the homeostatic phase}

\item{time.max}{maximal simulation time}

\item{time.samples}{time points at which simulation results are stored}

\item{N}{number of stem cells during homeostasis}

\item{NT1}{number of type 1 cells during homeostasis}

\item{NT2}{number of type 2 cells during homeostasis}

\item{report.at.f}{frequency of a selected clone at which the simulation should be stopped. Defaults to NA - don't stop at a certain frequency.}

\item{mut.rate}{average number of mutation per division and daughter cell}

\item{mutation.mode}{should mutations be "constant" or "Binomial"ly distributed?}

\item{driver.mode}{string; "fixed_time" if driver is acquired at a fixed time point, "random", if driver is randomly acquired governed by the parameters mu_D and s}

\item{t.driver}{integer vector, the time point(s) at which the driver is acquired if \code{mode} is "fixed_time"}

\item{mother}{integer vector containing the mother clone of each daughter clone; will be ignored if \code{driver.mode} is "random".}

\item{mut.rate.D}{integer; driver mutation rate per cell division}

\item{s.shape}{shape parameter of the gamma distribution from which the selective advantage of a new driver is drawn; if \code{driver.mode}== "fixed_time" and \code{t.driver} is a vector of multiple events, \code{s.shape} must be a vector of equal length as \code{t.driver}}

\item{s.rate}{rate parameter of the gamma distribution from which the selective advantage of a new driver is drawn; if \code{driver.mode}== "fixed_time" and \code{t.driver} is a vector of multiple events, \code{s.shape} must be a vector of equal length as \code{t.driver}}

\item{tau}{tau leaping parameter: how many steps should be merged during homeostasis? Defaults to 1, no leaping.}
}
\value{
a list of state.lists at the desired time samples
}
\description{
Simulate the phylogenetic tree of a physiological population that grows in 2 regimes: initial exponential expansion and subsequent homeostasis. The function can simuate stem cells only or stem and progenitor cells
}
