% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{histogram.drift}
\alias{histogram.drift}
\title{Computes the expected histogram of variants in a clone of interest at time t, given a VAF histogram at time zero}
\usage{
histogram.drift(
  lower.bins.1,
  n.muts,
  bin.p1 = 1,
  bin.p2,
  lower.bins.2,
  N,
  lambda,
  delta,
  t
)
}
\arguments{
\item{lower.bins.1}{vector of bin sizes of the histogram at t0}

\item{n.muts}{vector of mutation counts per bin at t0}

\item{bin.p1}{vector of probabilities that the variants in each bin will remain in the clone of interest}

\item{bin.p2}{as \code{bin.p1} but for the upper border of the bin}

\item{lower.bins.2}{vector of clone sizes at t}

\item{N}{the number of cells in the system}

\item{lambda}{the division rate}

\item{delta}{the loss rate}

\item{t}{the time point of evaluation}
}
\value{
The cumulative VAF histogram at time t
}
\description{
Computes the expected histogram of variants in a clone of interest at time t, given a VAF histogram at time zero
}
