% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igblast.R
\name{igblast}
\alias{igblast}
\title{Run IgDiscover for IgBlast using reticulate and conda environment}
\usage{
igblast(database = "path/to/folder", fasta = "path/to/file", threads = 1)
}
\arguments{
\item{database}{Vector containing the database for VDJ sequences}

\item{fasta}{Vector containing the sequences, usually a column from a data.frame. eg. df$sequences}
}
\value{
Creates a data frame with the Igblast analysis where each row is the tested sequence with columns containing the results for each sequence
}
\description{
Run IgDiscover for IgBlast using reticulate and conda environment
}
\examples{
## Example with test sequences
igblast(
    database = system.file("inst/extdata/test_fasta/KIMDB_rm"),
                                ,
    fasta = system.file("inst/extdata/test_fasta/test_igblast.txt")
)
}
