% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{mutational.burden.general}
\alias{mutational.burden.general}
\title{Mutation accumulation during exponential expansion followed by second phase of either expansion, homeostasis or decline}
\usage{
mutational.burden.general(
  mu,
  N.1,
  lambda.1,
  delta.1,
  lambda.2,
  delta.2 = NULL,
  t.end,
  b,
  accuracy.a = 0.05,
  phase = "both"
)
}
\arguments{
\item{mu}{mutation rate per cell division}

\item{N.1}{population size after first phase}

\item{lambda.1}{proliferation rate during initial expansion}

\item{delta.1}{loss rate during initial expansion}

\item{lambda.2}{proliferation rate during second phase}

\item{delta.2}{loss rate during second phase}

\item{t.end}{end point (starting after initial expansion)}

\item{b}{minimal clone size of interest. Number or vector.}

\item{accuracy.a}{step size in which mutations accumulated during expansion are evaluated (evaluation runs between 5 and 100\\%); defaults to 5\\%}

\item{phase}{return variants from "both" phases, or from "first" or "second" only}
}
\value{
This function returns the approximate number of mutations in clones of at least \code{b} cells, by first computing the distribution at the transition time within intervals, then averaging the fate of each interval during homeostasis and adding newly acquired mutations.
}
\description{
Mutation accumulation during exponential expansion followed by second phase of either expansion, homeostasis or decline
}
