% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution (conflicted copy
%   2023-06-16 181130).R
\name{mutational.burden.multiple.clones}
\alias{mutational.burden.multiple.clones}
\title{Mutation accumulation during exponential expansion followed by homeostasis with complex subclonal evolution.}
\usage{
mutational.burden.multiple.clones(
  mu,
  N,
  lambda.exp,
  delta.exp,
  lambda.ss,
  t.end,
  t.s,
  s,
  ancestry,
  b,
  min.clone.size = 0.05
)
}
\arguments{
\item{mu}{mutation rate per cell division}

\item{N}{population size}

\item{lambda.exp}{proliferation rate during expansion}

\item{delta.exp}{loss rate during expansion}

\item{lambda.ss}{proliferation and loss rate during homeostasis}

\item{t.end}{end point (starting from homeostasis)}

\item{t.s}{named vector of clonal birth dates}

\item{s}{named vector of selective advantages}

\item{b}{minimal clone size of interest. Number or vector.}

\item{tree}{an object of class phylo, giving the clonal relationships in the samples. Node labels must match names of t.s and s}
}
\value{
This function returns an approximation by first computing the distribution at the transition time within intervals, then averaging the fate of each interval during homeostasis and adding newly acquired mutations in a scenario with 2 nested clonal selections (clone starts growing at t.s >t.ss and grows with a selective advantage s; clone 2 starts. Returns the number of mutations present in at least b cells
}
\description{
Mutation accumulation during exponential expansion followed by homeostasis with complex subclonal evolution.
}
