% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{mutational.burden}
\alias{mutational.burden}
\title{Mutation accumulation during epxonential expansion followed by homeostasis.}
\usage{
mutational.burden(mu, N, lambda.exp, delta.exp, lambda.ss, t.end, b)
}
\arguments{
\item{mu}{mutation rate per cell division}

\item{N}{population size}

\item{lambda.exp}{proliferation rate during expansion}

\item{delta.exp}{loss rate during expansion}

\item{lambda.ss}{proliferation and loss rate during homeostasis}

\item{t.end}{end point (starting from homeostasis)}

\item{b}{minimal clone size of interest. Number or vector.}

\item{return}{This function returns an approximation by first computing the distribution at the transition time within intervals, then averaging the fate of each interval during homeostasis and adding newly acquired mutations. Returns the number of mutations present in at least b cells.}
}
\description{
Mutation accumulation during epxonential expansion followed by homeostasis.
}
