% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{mutational.burden.with.nested.selection}
\alias{mutational.burden.with.nested.selection}
\title{Mutation accumulation during exponential expansion followed by homeostasis.}
\usage{
mutational.burden.with.nested.selection(
  mu,
  N,
  lambda.exp,
  delta.exp,
  lambda.ss,
  t.end,
  t.s,
  t.s2,
  s1,
  s2,
  b
)
}
\arguments{
\item{mu}{mutation rate per cell division}

\item{N}{population size}

\item{lambda.exp}{proliferation rate during expansion}

\item{delta.exp}{loss rate during expansion}

\item{lambda.ss}{proliferation and loss rate during homeostasis}

\item{t.end}{end point (starting from homeostasis)}

\item{t.s}{time point at which selective advantage is acquired.}

\item{t.s2}{time point of second mutation}

\item{s1}{selective advantage associated with first mutation}

\item{s2}{selective advantage associated with second mutation}

\item{b}{minimal clone size of interest. Number or vector.}
}
\value{
This function returns an approximation by first computing the distribution at the transition time within intervals, then averaging the fate of each interval during homeostasis and adding newly acquired mutations in a scenario with 2 nested clonal selections (clone starts growing at t.s >t.ss and grows with a selective advantage s; clone 2 starts. Returns the number of mutations present in at least b cells
}
\description{
Mutation accumulation during exponential expansion followed by homeostasis.
}
