% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{mutational.burden.with.selection.no.size.compensation}
\alias{mutational.burden.with.selection.no.size.compensation}
\title{Mutation accumulation during exponential expansion followed by a second phase with clonal selection.}
\usage{
mutational.burden.with.selection.no.size.compensation(
  mu,
  N,
  lambda.exp,
  delta.exp,
  lambda.ss,
  t.end,
  t.s,
  s,
  b,
  accuracy.a = 0.05,
  min.clone.size = 0.05
)
}
\arguments{
\item{mu}{mutation rate per cell division}

\item{N}{population size during homeostasis in absence of CH}

\item{lambda.exp}{proliferation rate during expansion}

\item{delta.exp}{loss rate during expansion}

\item{lambda.ss}{proliferation and loss rate during homeostasis}

\item{t.end}{end point (starting from homeostasis)}

\item{t.s}{time point at which selective advantage is acquired.}

\item{s}{selective advantage}

\item{b}{minimal clone size of interest. Number or vector.}

\item{accuracy.a}{step size in which mutations accumulated during expansion are evaluated between 5 and 100\\%; defaults to 5\\%}

\item{min.clone.size}{the lower detection limit for selected clones}
}
\value{
This function returns an approximation by first computing the distribution at the transition time within intervals, then averaging the fate of each interval during homeostasis and adding newly acquired mutations in a scenario where a subpopulation is under positive selection. Returns the number of mutations present in at least \code{b} cells.
}
\description{
Mutation accumulation during exponential expansion followed by a second phase with clonal selection.
}
