% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{mutations.during.steady.state}
\alias{mutations.during.steady.state}
\title{Neutral mutation accumulation during steady state.}
\usage{
mutations.during.steady.state(lambda, N, mu, n.min, t.end)
}
\arguments{
\item{lambda}{proliferation rate}

\item{N}{population size}

\item{mu}{mutation rate per cell division}

\item{n.min}{minimal clone size}

\item{t.end}{time at end point}
}
\value{
The number of mutations that were acquired during steady state and are present in at least n.min cells.
}
\description{
Neutral mutation accumulation during steady state.
}
