% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{p.ss}
\alias{p.ss}
\title{Clone size distribution in a critical b-d process.}
\usage{
p.ss(lambda, a, b, t)
}
\arguments{
\item{lambda}{proliferation rate}

\item{a}{clone size at \code{t=0}}

\item{b}{clone size at \code{t=t}; single value or vector}

\item{t}{time}
}
\value{
The probability to grow from \code{a} to \code{b} within \code{t}
}
\description{
Function to compute the probability to grow from \code{a} to \code{b} in a critical b-d process using automatic switching between exact and approximate solution.
}
\details{
The function automatically switches between the exact solution and an approximation with a parametrized \eqn{\Gamma}-distribution at a cutoff criterion of \eqn{a*p*(1-p)>=9 \& b*p*(1-p)>=9}
}
