% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{probability.this.combination}
\alias{probability.this.combination}
\title{Probability of a variant present in a given bin size ends up in a particular combination of selected daughters if a driver is acquired in a random cell of the mother clone}
\usage{
probability.this.combination(
  bin.size,
  clone.size.mother,
  n.daughters.present,
  n.daughters.absent
)
}
\arguments{
\item{bin.size}{vector of bin sizes the variant are present in}

\item{clone.size.mother}{the size of the mother clone}

\item{n.daughters.present}{the number of daughters the variant ends up in}

\item{n.daughters.absent}{the number of daughters the variant does not end up in}
}
\value{
Computes the probability that a given variant that is present in \code{bin.size} cells of the mother clone ends up in \code{n.daughters.present} daughter clones, but not in the remaining \code{n.daughters.absent} if the cells giving rise to the selected daughters are randomly sampled.
}
\description{
Probability of a variant present in a given bin size ends up in a particular combination of selected daughters if a driver is acquired in a random cell of the mother clone
}
