% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.report.R
\name{quality.report}
\alias{quality.report}
\title{Generate general and individualized reports}
\usage{
quality.report(data_folder, outputfile, output_dir, processors = 1)
}
\arguments{
\item{data_folder}{Full file directory for searching all ab1 files in a recursive search method (all files including files in subfolders)}

\item{outputfile}{Output file name for the report generation}

\item{output_dir}{Output directory for all the different output files that are generated during the report}
}
\value{
HTML reports, fasta files, csv file
}
\description{
This function uses the other functions already described to create a HTML report based on sequency quality. Besides the HTML reports, it also creates fasta files with all the sequences and individualized sequences, in addition to a csv file with the quality scores and sequences considered as good quality.
}
