% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{quality_report}
\alias{quality_report}
\title{Generate general and individualized reports}
\usage{
quality_report(
  data_folder,
  outputfile = "QC_report.html",
  output_dir = "test/",
  processors = 1,
  folder_path_fcs = "path/to/fcs_datasets",
  plot_chromatogram = FALSE
)
}
\arguments{
\item{data_folder}{Full file directory for searching all ab1 files in a recursive search method (all files including files in subfolders)}

\item{outputfile}{Output file name for the report generation}

\item{output_dir}{Output directory for all the different output files that are generated during the report}

\item{processors}{Number of processors to use, or NULL (the default) for all available processors}

\item{folder_path_fcs}{Full file directory for searching all flow cytometry index files (.fcs) in a recursive search method (all files including files in subfolders)}

\item{plot_chromatogram}{Logical argument, TRUE or FALSE, to indicate if chromatograms should be plotted or not. Default is FALSE.}
}
\value{
HTML reports, fasta files, csv file
}
\description{
This function uses the other functions already described to create a HTML report based on sequency quality. Besides the HTML reports, it also creates fasta files with all the sequences and individualized sequences, in addition to a csv file with the quality scores and sequences considered as good quality.
}
\examples{
\dontrun{
quality.report(
  data_folder = "test/test_dataset/sanger_sequences",
  outputfile = "QC-report.html",
  output_dir = "test/",
  processors = 1
)
}
}
