% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepertoiR.R
\docType{package}
\name{RepertoiR}
\alias{RepertoiR}
\title{B cell receptor immune repertoire quality control and analysis}
\description{
RepertoiR was created to do quality control of VDJ sequencing,
combining single-cell sorted data (FACS) and specificity with sanguer sequence information.
}
\author{
Rodrigo Arcoverde Cerveira \email{rodrigo.arcoverdi@gmail.com}
}
