% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scifer.R
\docType{package}
\name{scifer}
\alias{scifer}
\title{scifer: Single-cell Immunoglobulin Filtering of Sanger sequences}
\description{
Integrating index single-cell sorted files with Sanger sequencing per plates,
combining single-cell sorted data (FACS) and specificity with Sanger sequencing information.
}
\author{
Rodrigo Arcoverde Cerveira \email{rodrigo.arcoverdi@gmail.com}
}
