% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_post_processing.R
\name{simulate_vaf_upon_sequencing}
\alias{simulate_vaf_upon_sequencing}
\title{Simulate the measured VAF after sequencing}
\usage{
simulate_vaf_upon_sequencing(vaf, depth)
}
\arguments{
\item{vaf}{a vector of true VAFs in the population}

\item{depth}{average coverage in sequencing}
}
\value{
a vector of simulated VAFs after sequencing
}
\description{
Simulate the measured VAF after sequencing
}
