% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{simulated.data}
\alias{simulated.data}
\title{Wrapper function to simulate sequencing either by bulk WGS or by scWGS}
\usage{
simulated.data(
  seqtype,
  clone.sizes,
  expected.mutations,
  depth = 90,
  ncells = 100,
  sensitivity = T,
  false.negative.per.vaf,
  min.vaf = 0.05
)
}
\arguments{
\item{seqtype}{string, specifying the sequencing method. Must be either 'bulk' or 'sc'}

\item{clone.sizes}{vector of clone sizes at which cumulative mutation counts were measured}

\item{expected.mutations}{expected number of mutations at each clone size}

\item{depth}{sequencing depth, only specify for bulk WGS}

\item{ncells}{the number of sequenced cells, only specify if seqtype="sc"}

\item{sensitivity}{logical, if sensitivty of sequencing method should be taken into account in addition to binomial noise. Requires a specification for false.negative.per.vaf}

\item{false.negative.per.vaf}{matrix with columns corresponding to the measured VAFs and rows corresponding to individual measurements of the false negative rate at this VAF.}

\item{min.vaf}{the minimal VAF to return}
}
\value{
A vector of simulated VAFs
}
\description{
Wrapper function to simulate sequencing either by bulk WGS or by scWGS
}
