% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{.clonal_dynamics}
\alias{.clonal_dynamics}
\title{Dynamics of normal cells and j selected clones}
\usage{
.clonal_dynamics(N, init, lambda, delta, s, t)
}
\arguments{
\item{N}{the carrying capacity}

\item{init}{vector of length j with the initial condition of the system (number of cells per clone)}

\item{lambda}{cell division rate}

\item{delta}{differentiation rate}

\item{s}{vector with selective advantage associated with the j-th driver. Selection is modeled as a reduction of the differentiation rate, so 0 <= s <= 1}

\item{t}{the time point of evaluation}
}
\value{
The system state at time t
}
\description{
Dynamics of normal cells and j selected clones
}
