% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{.simulated.wgs.data}
\alias{.simulated.wgs.data}
\title{Simulate read sampling in WGS by Binomial sampling}
\usage{
.simulated.wgs.data(
  clone.sizes,
  expected.mutations,
  depth = 90,
  sensitivity = T,
  false.negative.per.vaf,
  min.vaf = 0.05
)
}
\arguments{
\item{clone.sizes}{vector of clone sizes at which cumulative mutation counts were measured}

\item{expected.mutations}{expected number of mutations at each clone size}

\item{depth}{sequencing depth}

\item{sensitivity}{logical, if sensitivity of sequencing method should be taken into account in addition to binomial noise. Requires a specification for \code{false.negative.per.vaf}.}

\item{false.negative.per.vaf}{optional, a matrix with columns corresponding to the measured VAFs and rows corresponding to individual measurements of the false negative rate at this VAF in addition to binomial noise. Must be provided if \code{sensitivity=T}}
}
\value{
A vector of simulated VAFs
}
\description{
Simulate read sampling in WGS by Binomial sampling
}
