% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_info_from_vcf.R
\name{Extract.info.from.vcf}
\alias{Extract.info.from.vcf}
\title{Extracts information from a vcf file}
\usage{
Extract.info.from.vcf(
  vcf,
  info = "readcounts",
  type = "snvs",
  mutationcaller = "Strelka",
  tumor.col.mutect = 10,
  normal.col.mutect = 11,
  sample.col.mpileup = NA
)
}
\arguments{
\item{vcf}{Mutation file in VCF representation as a list (as returned by read.vcf from package bedR).}

\item{info}{Which information is to be retrieved? Possible values are "readcounts", "depth", "VAF", "AA_change", "cDNA_change", "Exon", "VAF.control", "depth.control", "Gene", "annovar_function", "exonic_function". Defaults to "readcounts".}

\item{type}{Is the VCF file reporting snvs or indels? Possible values are "snvs" and "indel". Defaults to "snvs".}

\item{mutationcaller}{which mutation caller was used to generate the vcf-files? Defaults to "Strelka".}
}
\value{
purity and ploidy
}
\description{
Extracts information from a vcf file
}
