% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{mutational.burden}
\alias{mutational.burden}
\title{Mutation accumulation during exponential expansion followed by homeostasis.}
\usage{
mutational.burden(
  mu,
  N,
  lambda.exp,
  delta.exp,
  lambda.ss,
  t.end,
  b,
  accuracy.a = 0.05,
  phase = "both"
)
}
\arguments{
\item{mu}{mutation rate per cell division}

\item{N}{population size}

\item{lambda.exp}{proliferation rate during expansion}

\item{delta.exp}{loss rate during expansion}

\item{lambda.ss}{proliferation and loss rate during homeostasis}

\item{t.end}{end point (starting from homeostasis)}

\item{b}{minimal clone size of interest. Number or vector.}

\item{accuracy.a}{step size in which mutations accumulated during expansion are evaluated (evaluation runs between 5 and 100\\%); defaults to 5\\%}

\item{phase}{return variants from "both" phases, or from "expansion" or "homeostasis" only}
}
\value{
This function returns the approximate number of mutations in clones of at least \code{b} cells, by first computing the distribution at the transition time within intervals, then averaging the fate of each interval during homeostasis and adding newly acquired mutations.
}
\description{
Mutation accumulation during exponential expansion followed by homeostasis.
}
