% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution (conflicted copy
%   2023-06-16 181130).R, R/Clone_size_distribution.R
\name{p.ss.approx}
\alias{p.ss.approx}
\title{Approximate solution to grow from a clone of size a to a clone of size b within t in a critical birth-death process using gamma distribution}
\usage{
p.ss.approx(lambda, a, b, t, mode = "density")

p.ss.approx(lambda, a, b, t, mode = "density")
}
\arguments{
\item{lambda}{proliferation rate}

\item{a}{clone size at t=0}

\item{b}{clone size at t=t}

\item{t}{time}

\item{mode}{either 'density' if density distribution is to be returned or 'cumulative'.}
}
\value{
The approximate probability to grow from a to b within t.

The approximate probability to grow from a to b within t.
}
\description{
Probabilitly to grow from a clone of size a to a clone of size b within t in a critical birth-death process using a parametrized gamma distribution for approximation.
}
