% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution.R
\name{p.ss.exact}
\alias{p.ss.exact}
\title{Clone size distribution of a critical birth-death process (exact).}
\usage{
p.ss.exact(lambda, a, b, t)
}
\arguments{
\item{lambda}{proliferation rate}

\item{a}{clone size at t=0}

\item{b}{clone size at t=t}

\item{t}{time}
}
\value{
The probability to grow from a to b within t.
}
\description{
Exact solution to grow from a clone of size a to a clone of size b within t in a critical birth-death process.
}
