% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clone_size_distribution (conflicted copy
%   2023-06-16 181130).R, R/Clone_size_distribution.R
\name{p.ss}
\alias{p.ss}
\title{Function to compute the probabilty to grow from a to b in a critical b-d process using automatic switching between exact and approximate solution.
Cutoff criterion: a*p*(1-p)>=9 and b*p*(1-p)>=9}
\usage{
p.ss(lambda, a, b, t)

p.ss(lambda, a, b, t)
}
\arguments{
\item{lambda}{proliferation rate}

\item{a}{clone size at t=0}

\item{b}{clone size at t=t; single value or vector}

\item{t}{time}

\item{mode}{either 'density' if density distribution is to be returned or 'cumulative'.}
}
\value{
The probability to grow from a to b within t

The probability to grow from a to b within t
}
\description{
Function to compute the probabilty to grow from a to b in a critical b-d process using automatic switching between exact and approximate solution. Cutoff criterion: a*p*(1-p)>=9 and b*p*(1-p)>=9
}
