% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scifer.R
\docType{package}
\name{scifer}
\alias{scifer}
\title{Scifer: Single-Cell Immunoglobulin Filtering of Sanger Sequences}
\description{
Integrating index single-cell sorted files with Sanger sequencing per plates,
combining single-cell sorted data (FACS) and specificity with Sanger sequencing information.
}
\author{
Rodrigo Arcoverde Cerveira \email{rodrigo.arcoverdi@gmail.com}
}
\keyword{AIRR}
\keyword{BCR}
\keyword{QC}
\keyword{TCR}
\keyword{flowcytometry}
\keyword{sanger}
\keyword{sequencing}
